+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - none
 
 # v2022.10.1-beta
+- demo - add chart tradebot_historical_backtrack
 - chart - allow hiding series by clicking it in plot-area
 - coverage - add test-coverage for function sqlmathWebworkerInit()
 - chart - use different color for category - short